Update CPGQL terminology
Some of the wording used to describe various CPGQL steps was inconsistent and also not flexible enough to describe dataflow steps. This change in terminology fixes those two issues by making the `Directive` a basic building block for definitions in CPGQL.

```
> OVERVIEW

>>> Directive
CPGQL Directives are keywords of the Code Property Graph Query Language.

>>> Step
CPGQL Steps are combinations of one or more Directives that describe graph traversals in the CPGQL. They are represented by one or more Directives.

>>> Node-Type Step
CPGQL Node-Type Steps are Steps that traverse nodes based on their type. They are represented by a single Directive.

>>> Filter Step
CPGQL Filter Steps are Steps which filter nodes in a traversal according to a criterion. They are represented by one or more Directives.

>>> Repeat Step
CPGQL Repeat Steps are Steps which repeat another traversal multiple times. They are represented by one or more Directives.

>>> Complex Step
CPGQL Complex Steps are Step which combine the functionality of one or more Node-Type Step, Repeat Step, Filter Step, Core Step or Execution Directive. They are represented by one or more Directives.

>>> Core Step
CPGQL Core Steps are Steps which can be combined with any other Step. They are represented by one or more Directives.

>>> Execution Directive
CPGQL Execution Directives are Directives which suffix Steps and execute the traversal the Steps represent.

>>> Augmentation Directive
CPGQL Augmentation Directives are Directives which extend a Code Property Graph with nodes, properties and edges.

>>> Help Directive
The CPGQL Help Directive is a Directive which returns textual descriptions of other directives.

>>> Entry Directive
A CPGQL Entry Directive is a Directive which references the entry node of a Code Property Graph.

>>> Query
A CPGQL Query is a combination of more than two Directives.
```