Improve column calculation algorithm.
The old version managed to be slow enough to be the number one hot spot
during profiling of c2cpg. The new version makes this fast so that it
does not show up during performance measuring anymore.

Column number now start at one instead of zero like the line numbers and
the new implemention obsoleted a bug in the old column end calculation
which led to complete bogus numbers.