jssrc2cpg: bring in native astgen invocation (#1617)
I was finally able to build a native custom version of astgen (based on @joernio/astgen 2.0.4) using 'pkg'.
This enables us to run it entirely without node (as a stripped down node comes with the binary), and it is ultimately much faster.  Updated README w.r.t. the binary build process.