jssrc2cpg: removed test resource files (#1641)
Test files and their content that is required during jssrc2cpg tests is now generated on-the-fly.
This is way easier to maintain as the single point of truth is now directly visible in the test code.
Hence, we do not have to rely on external files any longer. Also, the handling of resource paths / URIs during testing within Intellij or a SBT shell is quite error prone.