Fixed file reading in presence of unknown line separators
Input files may have been created on different platforms with different tools.
Hence, we might end up with a mixture of line separation characters across files and can not simply
assume \n is safe to append when trying to calculate code fields or line/column numbers.

This is especially problematic within jssrc2cpg as it relies on the correct mapping
between the actual source file and the JSON generated by astgen for calculating code fields and line/column numbers.