Simplifiy FlowGraph trait. (#1979)
* Simplifiy FlowGraph trait.

Remove entryNode and exitNode. Those are not required by
DataFlowProblem and DataFlowSolver and where so far only used
to piggy back the method node onto the data flow problem for the
using codes sake.
Now the using code handles the method node itself and on top we avoid
casting.

* Made succ and pred functions.

This enables implementors to choose how they want to implement those
to relations and are not forced to use a Map.