WIP add (failing) test to verify that two calls are part of the DDG
I'm not the best person to judge wether those calls really _should_
be part of the DDG, but from our discussion on
[discord](https://discord.com/channels/832209896089976854/832209896089976857/1044687371800678420)
it sounds like the answer is 'yes'.
Anywhere, since I was asked about this I just added a (failing) test
to hopefully make it easier to grok the context and take it from
there.