CodeDumper: safer filename calculation (#2094)
For methods generated by the MethodStubCreator (e.g., for operators) we end up with artificial method nodes
in the <global> namespace and file but without line numbers.
The fake filename (<global>) would crash the filename calculation here. So I moved that
piece of code where it is guarded by lineNumber existence.