Xmx max heap memory handling for importing large codebases (#2140)
Prior to this, we would not pass the Xmx (max heap) setting from joern to frontends.
Only a globally configured `JAVA_OPTS` would work, but most users would either configure
that specifically for joern (e.g. `JAVA_OPTS=-Xmx30g ./joern`) or use the `-J` flag,
e.g. `./joern -J-Xmx30g` - especially if they want to import large codebases. 

Now, we determine the maxMemory that the user configured joern with, and pass that to 
all JVM based frontends. Since the sum of joern and the frontend may exceed the physical
memory of the machine, we tell the user that they may want to exit joern and invoke the 
frontend themselves.  Determining the available max memory for a machine is tricky, so 
we'll keep it simple for now.