fix logging dependencies setup (#2228)
* fix logging dependencies setup

We declare slf4j-api as a regular dependency, since most logging
libraries have an implementation for it, and it is considered
a lowest common denominator for logging.

We choose log4j as the logging provider for this build, i.e. for tests
and `sbt stage`. `Optional` means "not transitive", but still included
in "stage/lib". I.e. downstream builds do not inherit it and can/must
make their own choices.

* fixup ghidra2cpg mess - log4j is included in the fat jar we depend on :(

* latest cpg