Increase Memory Alloc (#2855)
* Increase Memory Alloc and Constrain Test Tasks
Many of the runners on the CI/CD are aborting the test suites on occasion. This is likely to the scaling number of tests and parallelism of the test execution demanding more memory as development continues.

Setting the task limit on tests to be 2 x available cores seems to crash the Windows runner so I've just updated the memory via `.sbtopts`.