Make sure `lone identifier optimization` respects `return`s (#2873)
We previously removed lone identifiers even if they actually also appeared in `return` statements. That was incorrect but not triggered when the affected identifier was a local or parameter. The bug became apparent because in the ruby frontend, locals are currently missing. While we should add locals to the ruby CPG, this fix should be made nonetheless and it fixes a missing flow we were observing.


Co-authored-by: Fabian Yamaguchi <fabs@joern.io>