Add consistent recursive jar unpacking.
Previously jimple2cpg would have two distinct
code paths when importing a directory containing a
jar (which would unpack the jar) compared to importing
a jar directly (which would load it with soot).

This change does the following
- Always use the unpacking behaviour for jars
- Remove jimple from the source files being processed
  (It can be added back if needed downstream)
- Keep the .apk/.dex behaviour if imported directly (but not if contained
in a directory). This maintains the existing behaviour.

Some remaining work is to:
- add docstrings
- add unit tests for recursive unpacking
- add unit tests for .apk/.dex (maybe)