[TypeRecovery] Bound Symbol Table Value Set Following [this thread](https://github.com/joernio/joern/issues/3211#issuecomment-1665063603), @yzn12345 found that the symbol table value sets could grow to extremely large sizes, as much of the "dummy" values simulate possible access paths at an exponential rate of sorts.
The purpose of these dummy types is simply to indicate where a type is coming from, but much of these N types share the same prefix though end in some large number of suffix variants which aren't necessary.

This PR coalesce's new additions, prioritizing non-dummy types, but limiting the value set at 10. This shows a huge increase in speedup and memory efficiency.