Code Refactor
Did following code refactor

1. Remove ThreadLocal for `fullyQualifiedPackage` as it was unnecessary
in the first place (For every file being processed a new AstCreator
object is being created)
2. Did some refactoring of code to move some constants at common place.
3. Used common `getTypeFullName` method while constructing `ParameterIn`
method nodes.
4. Refactored Import nodes to consider last path element of the import
node as default alias and set it for `importedAs`. As this is how its
been treated in `Go`. Along this change updated respective unit tests
for the same.