[CfgCreator] Fixed fringe erasure in try-blocks (#3587)
In the case of a try-block, if an explicit return node was given then the fringe would be erased and the Cfg would no longer be a graph rooted in the method node. The first statement in the try block would have it's fringe erased by the Cfg of the return node and thus be dangling.

This change would detect if the return node is within a try-block and conditionally inherit the fringe.