Initial code refactor to build the cache
While resolving the the methodfullnames and typefullname for chained
method calls relised that we need to have the cache of `methodFullName`
mapped to (returnType, signture), before we start processing the AST.
This will help identifying at least inernal source call node
TypeFullNames and signatures for all possible combinations.

1. Created new pass which will run through all the AST files and only
process `ast.FuncDecl` nodes to build this cache. This has been done by
extending AstCreator as CacheBuilder. Preserving the information that we
already processed and use the same AstCreator while running the
AstCreationPass.
2. Refactored the code acording to the above mentioned restructuring in
order to reuse the code.