Fix for issue #3678
Fix for issue #3678.

For now this PR fixes the issue related to breaking error for Local node
being pushed to as condition AST node for switch AST. Fixed the issue
and resolved the breaking error.

This will still require a proper fix to push the Assign AST as first
child of which block AST. We will handle this issue at later stage.