package name is different from namespace folder name
1. Handling for the scenario when the package name and namespace folder
name are different.
In the below example, we are importing `joern.io/sample/lib` but `fpkg`
alias is getting used. We created a global cache for this mapping and
did the lookup if the alias to namespace lookup fails for file level
mapping we maintain.
```
package main
import "joern.io/sample/lib"
func main() {
	var a = fpkg.Woo(10)
	var b = fpkg.Sample{name: "Pandurang"}
	var c = b.Name
	var d fpkg.Sample
}
```