[gosrc2cpg] - Handling When Package Name is Different to Enclosing Folder of Package (#3699)
**package name is different from the namespace folder name**
Handling for the scenario when the package name and namespace folder name are different.
In the below example, we are importing `joern.io/sample/lib` but `fpkg` alias is getting used. We created a global cache for this mapping and did the lookup if the alias to namespace lookup fails for the file level mapping we maintain.

```
package main
import "joern.io/sample/lib"
func main() {
	var a = fpkg.Woo(10)
	var b = fpkg.Sample{name: "Pandurang"}
	var c = b.Name
	var d fpkg.Sample
}
```
**Type declared after its being used**

Member Type cache wasn't getting populated before AST Creation pass is invoked. It seems the respective unit test didn't cover the scenario, which will cover this scenario where it is required to have a cache built before AST creation pass.

Added respective unit tests along with the fix for the same.