[gosrc2cpg] - Made structural changes while creating initial AST nodes per file #3695 (#3734)
Earlier we were creating AST structure per file in following way.

NameSpaceNode(`package prefixed with filename`)=>TypeDecl(`Fake TypeDecl
node representing the given file`)=>Method(`Fake Method node
representing the statements (imports, global var and constants, types and
functions) in the given file`)=>Block(`all the direct statement ASTs as
child nodes`)

With new structure, we wanted to create Fake TypeDecl node representing
one package (One package can have multiple go files in it). In order to
achieve this, we are following order of nodes being created for package
and then for a file.

For package, i.e. per folder

NameSpaceNode(`one namespace node representing one
package`)=>TypeDecl(`Fake TypeDecl node representing the give package.
File path mentioned in this case will be the folder path`)

For each file in the package. The first fake node will be added as child
node of the Package TypeDecl created in the previous cache pass.

Method(`Fake Method node representing the statements (imports, global var
and constants, types and functions) in the given file`)=>Block(`all the
direct statement asts as child nodes`)

- Added few more unit tests covering the new use cases
- Changed the existing unit tests to adjust with updated AST structure.
- A few memory optimisations to exclude some part of processing while
processing third party dependencies

TODO:

- Handle the global variable and constants to be created as member
fields of the package level Fake TypeDecl we are creating. As of now it's
being created as `LOCAL` node under Fake file level `METHOD` node we have
created.
- If the Global variable or constant is getting accessed in the same
package. As of now that is being treated as an `IDENTIFIER`, we need
change to be converted to FieldAccess `CALL` node.
- With the above change, we might not require the File level fake
`METHOD` node as for Global variable or constants we are going to treat
them as Member fields of level Type Decl instead of `LOCAL` nodes. So we
might get rid of this file-level Fake `METHOD` node.