disambiguate implicits for things like MethodParmeterIn.method
Complicated stuff... let's use MethodParameterIn as an example. The
schema defines `.method` as a step which (for flatgraph) leads to the
generation of an extension step via an implicit.
Additionally, this build defines `.method` for `CfgNode` via extension
steps. This leads to an ambiguous implicit - while the priority for
the generated one is lower, the implicit for `CfgNode` is less precise
if we have a `MethodParameterIn` at hand.
To disambiguate, we need to move the conversions for the base node
types (e.g. CfgNode) into another (lower) mixin trait.

Note: previously `.method` was defined on the generated node types directly,
which avoided the ambiguous implicits we got ourselves into here.