[php2cpg] Php type recovery (#3723)
* Create PHP Type Recovery pass

* Remove erroneous <init> types from symbolTable

* Implement builtins type recovery lookup

* Update PhpTypeRecovery pass to lookup in CPG

* Remove overrides of methods that are meant to be private

* Run scalafmt

* Create initial test file for PhpTypeRecoveryPass

* Add PhpTypeRecoveryPassTests

* Add PhpSetKnownTypes pass for builtin PHP functions

Pass will set the types for builtin functions with known function
signatures. This pass will run before the PhpTypeRecovery pass.

Currently, the pass does not handle variadic parameters.

* Minor update to tests for SetKnownTypes pass

* Fix formatting with scalafmt

* WIP: Refactor PhpSetKnowTypes pass

* WIP: Refactor PhpSetKnownTypes pass to be cleaner

Currently the method return types are set correctly, but need to still
update the parameter types.

* Refactor PhpSetKnownTypesPass as a ForkJoinParallelCpgPass

* Ignore tests that depend on context sensitivity and comments

* Ignore more context sensitive tests

* Remove multi-line debug statements

* Remove builtins table from PhpTypeRecoveryPass

* Fix how multiple return statements are handled

This bug comes from the XTypeRecovery pass, but is fixed in
PhpTypeRecovery by overriding the visitReturns method. If this approach
(using another "symbol table" for methods) is sufficient, it should also
be fixed abstract parent class.

* Fix typo in PhpTypeRecoveryPassTests

* Filter out dummy return values when saving types

* Refactor the dummy type removal to be less coarse

* Resolve unknown namespaces with new type info

* Ignore array field type recovery tests for now

* Run scalafmt

* Fix MatchError in resolving dynamic calls

* Implement Php2Cpg.postProcessingPasses for deduplication

This defines the list of default PHP postProcessingPasses in one
location, so that they can all be applied without code duplication
between the different places where passes are applied, like
PhpCpgGenerator and PhpCode2CpgFixture.

Additionally, the XTypeRecoveryConfig options are now exposed to the
frontend command line arguments for PHP analysis as a result of this
refactor.

* Refactor PhpSetKnowTypes use idiomatic scala matches

* Replace hardcoded string with defined value

* Remove excessive debug statements and format comments

* Run test:scalafmt

* Address minor comments

- Changed string equality comparisons from equals() to ==
- Changed "<module>" to "<global>" as top-level method namespace
- Removed errant debug statement