fix for package-level global variables from dependencies.
1. It's been observed that the AST nodes were getting created for
package-level global variables and constants from the dependencies code.
As it was getting processed through the same ASTCreators code. Made
changes to handle this situation.
2. Also identified that the variables, constants, types and functions
from dependencies which starts with lower case letter will not be
referred from the main source code. Hence, I ignored the processing for
those. This will also reduce the memory footprint as we are not storing
that metadata in the cache. This will also speed up the processing
3. With a lesser memory footprint, unit tests with download dependencies
should work. Hence enabled those unit tests. These tests will be helpful
in identifying some breaking changes. Like the other ignore test in the
same file, was working before. With some recent changes, it has started
breaking that flow.