[go] - Ignore indirect dependencies (#3964)
1. `go.mod` also contains indirect dependencies, as we were processing
all the dpendencies to cache type, and method signature metadata. Even
with few direct dependencies it may endup having many indirect
dependencies. Which are generally not used in main code. Hence for now
ignoring these indirect dependencies so that it can reduce the time it
takes to download and process those dependencies.
2. Introduced `include-indirect-dependencies` flag in case someone wants
to try downloading and processing indirect dependencies.