[jssrc2cpg] Fixed crash on Windows
`filename` may already be an absolute path on Windows systems when running jssrc2cpg with an absolute path as input directory because node.js Babel works differently there (keeping that absolute path).
When calling `in / filename` we would crash on Windows because concatenating two absolute paths with better.files throws an exception then.