[swiftsrc2cpg] Added support for IfConfigDecls
We can't statically evaluate the conditions for these #if defs (they work pretty much the same is in C/C++).
(plus, we do not want to re-implement the Swift pre-processor)

So we now provide a `--define` flag (same as in c2cpg) to set the desired config and generate all the sub ASTs accordingly.
Obviously, this will only work for simple identifiers whereas Swift allows for arbitrary expression as `#if` condition.