[x2cpg] Program Summary & Typed Scope (#4131)
A few frontends each face the problems of:
* What types are available at this point in the program? How do I resolve a type alias to its full name?
* How do I deal with tracking possible method calls?
* How do I deal with conflicts if more than one type/method share the same name?
* Do I need to download my dependencies to provide the type meta-data?

This PR standardises the various structures found across frontends that try to deal with this by having a program summary and modified scope object.

The `ProgramSummary` gives enough bare-bones structure to describe the meta-data of high level interprocedurally scoped objects. The idea is that this is populated via a pre-parse before the `AstCreator` runs, and can be populated with pre-computed type stubs. The `TypedScope` trait allows one to keep track of which types are currently available at the current point of the program to help manage type references and aliases.

See the test for an illustration of this scoped type look-up behaviour.

The `TypeMap` and `CSharpScope` structures were pilot studies into this strategy and have proven effective. This structure tries to be fairly broad and flexible, and will be back ported into CSharpSrc and new implemented in the new Ruby frontend (useful for parenthesisless calls of methods imported into the scope).