Kotlin frontend performance improvements
1. Changed input directory list parameter to a single main repo path. We
were earlier passing a list of all directories in the repository. Which
was resulting into `KotlinToJVMBytecodeCompiler.INSTANCE.analyze` method
taking too much of time along with consuming too much of memory in the
process. Tried with the single path and it reduced the time as well as
memory footprint with the same results.
2. Made changes into warning logs to add the file path to get the
details of the file where we are facing the given problem.
3. Removed two assert statements and handled the situation gracefully.