[jssrc2cpg][swiftsrc2cpg] Safer offset calculation
In the presence of unmappable UTF8 characters we should not calculate offsets out of bounds.
That happened if an unmappable UTF8 character was replaced by something of different length compared to the original source code.
(when reading the file with `IOUtils.readEntireFile`)

(also a few improvements to swiftsrc2cpg AST generation)