SimpleCache: wipe old files
In order for the cache not to grow up too much in size, we want to clear old entries.
This cache is so simple though that we have no idea when a file was last recently used, so we bite the bullet
and simply delete all files older than 30 days.
(In theory one could change the api to allow for that, but that'd make the SimpleCache api less 'simple')