Kotlin frontend FieldAccess CALL node issue fixes.
1. As part of the performance issue troubleshooting exercise came across
the issue where the FieldAccess `CALL` Node was getting created for
chained field access e.g. a.b.c. While creating the fieldAccess `CALL`
node for `.c` it was creating CALL node with `methodFullName` set to
`<operator>.fieldAccess` and the name was getting set as `c` instead of
the same value as `methodFullName`. Because of this, it resulted in
creating many `METHOD` node stubs. In turn in StaticCodeLinker pass it
was linking all the FieldAccess `CALL` Nodes to all of these `METHOD`
stubs as the look-up was done over `methodFullName`. Added respective
unit tests for the same.
2. Made respective changes to fix this issue.