[swiftsrc2cpg] Fixed handling of extensions from multiple files
Swift extensions may be defined in the same file as the base type or in any number of different files. If we generate the ASTs file-parallel we may encounter such an extension before, after, or in between the generation of the AST for the base type. Hence, during AST generation its not clear whether a type has an extension defined somewhere. Just by looking at the name (and using a global cache) we might merge arbitrary typedecls even if they are unrelated in addition to potential races due to the parallel processing of all *.swift files (happens in sptests atm).

This PR creates additional typedecls for extensions (<extension> suffix) and gets rid of locking and the global typedecl map.