[c2cpg] Emit Catch ControlStructures (#4379)
With this change catches do not have to have order 2 for correct control flow generation. Multiple siblings with the same order are invalid.
The change in CfgCreator allows for backwards compatibility.
We fall back to order(2) for catch and order(3) for finally if no ControlStructure CATCH or ControlStructure FINALLY is found.