[c2cpg] Handle preprocessed files (#4466)
cpp/gcc/g++ offer the -E switch. The -E switch forces the compiler to stop after the preprocessing phase and spit out the result where all preprocessor directives are expanded or processed, all macros expanded, and all header files included. Typically, this result is stored as *.i file.

With this PR we process such files and exclude all the others with the exact same name if `--with-preprocessed-files` is used.

Fixes: https://github.com/joernio/joern/issues/4462