[dataflowengineoss] Passes/PassesNot Fix In a more naive time in my life, I wrote https://github.com/joernio/joern/pull/1423, but this did not *exactly* achieve what I wanted, since this is designed for sanitizers. This also took a set argument, which was a bit weird.
This fixes the above approach, by actually checking the path for the element in the traversal.

Note: For some calls, one might need to have something like `sink.reachableByFlows(source).passesNot(_.collectAll[Expression].inCall.name("sanitizer"))"` as the way the path elements are presented, the call may be skipped in languages such as Python.

Resolves #2438
Resolves #2175