[java] Resolve External Non-JDK Types with Type Arguments (#4527)
It's been known that certain types return as `ANY` when it's an external type with type arguments. However, the type would resolve when type arguments are removed.

This PR detects if the type to be resolved is a `ClassOrInterfaceType`, and if so, will only pass the type name to the type resolver, and resolve the type arguments separately. The type arguments are then placed in the `fullName` if `keepFullNames` is enabled.