[c2cpg] Improve full names and signatures for C++.
The C++ method full names did not include signatures which caused a lot
of methods to have the same full name which is not allowed. The
situation is not yet fully resolved because method/function declarations
independent from the implementation still created method stubs with full
name collisions. But it is a step in the right direction and the stubs
are easy to filter out via `isStub` step.

The call node generation has been reworked to accomodate this change.
Most noteable changes there:
- Full names where required contain signatures
- Signatures do not contain method names anymore
- Dispatch type is properly set if known.