Mulit module support
Earlier, if we pass the directory path which contains multiple go
modules. Processing was done with all the `.go` files mapped to single
`go.mod` file.

With this change, we have segregated the processing by first isloating
all the files mapped to respective `go.mod`. This will also make sure to
cleanup the memory footprint after every module is being processed.
However this will increase the processing when used with download
dependency as it will process all the `go.mod` files for identifying and
processing used dpendencies.