[ruby] Approximate Attribute Assignment (#4777)
With Ruby attribute assignments being calls, this can lead to costly data-flow tracking and these setters are rarely overridden with custom logic. This PR simplifies this model by representing the attribute assignment from a setter call to a direct field assignment. A similar adjustment is done to member getters.

Misc: Simplified a parser bug workaround and linked it to an issue.