[ruby] New `parser` Gem Parser Implementation (#5040)
Due to the performance and stability issues with a 2-phase LLK parser approach with ANTLR, this branch, developed over the past two weeks, introduces the usage of a [`parser`](https://github.com/whitequark/parser/tree/master) based approach wrapped in [`ruby_ast_gen`](https://github.com/joernio/ruby_ast_gen). This Gem is used by many popular Ruby linters and static analysis tools, and is compatible with Ruby versions across 1-3. 

We bundle `ruby_ast_gen` with the frontend under `resources`, and execute the Ruby code using JRuby to maximize portability. If the resource is detected to be in a JAR, `ruby_ast_gen` is written to a temporary directory, from where it is then executed.

Both JRuby and native Ruby runs of `ruby_ast_gen` far outperform the old ANTLR version, which would often have worst-case performance for some set of language constructs encountered.

Changes include:
* Remove ANTLR implementation
* Execute `ruby_ast_gen` and parse JSON output 
* Map the JSON output to the same intermediate AST nodes that ANTLR used to avoid changing `AstCreation` logic
* Rename the method name for executable strings (`%x{}`-style strings) to `<operator>.backticks` instead of `exec` as `exec` technically has different semantics
* Trim `.code` properties to 60 characters to avoid large `.code` properties. (this is done by `ruby_ast_gen`)
* Added `db/migration` to default ignored directories

Tests are largely kept the same, and `ruby:test` is green on `codescience` after some small `.code` and `.lineNumber` tweaks. See `dave/ruby/prep-for-ruby-parse-gem`.

The `astGenResourceTask` task in `build.sbt` is tied to `Compile / resourceGenerators` as the task fetches the zipped Ruby code and places it under `resources` to be packaged alongside the classes in the JAR.
---------

Co-authored-by: Andrei Dreyer <andrei@whirlylabs.com>
Co-authored-by: Andrei Dreyer <andrei.dreyer@adagintech.com>