[pysrc2cpg] Fix captured variable linking into class scopes.
The linking of captured variables was not yet working if variables from
outside of a class have been captured into methods of a class.
This was the case because of a possible name clashes between those
captured variables and variables inside of the class body methods.
To avoid such clashes captures variables get a `<captured>` prefix
inside the class body methods.
This solves a long standing issue where data flows via such captured
variables were not found.