fixup initialisation order of cpg generators
We trigger the post-processing directly if a cpg is loaded on startup
(e.g. `./joern cpg.bin`). In that case, the typeRecoveryConfig has not
been initialized. There's a reason why in Scala you have to explicitly
make an effort to start with `null`, because more often than not
you'll be missing out some random edge case... lazy val is the better
alternative

fixes https://github.com/joernio/joern/issues/4999 (again...)