fix json serializer: define only for nodes, not for all Products (#5612)
* fix json serializer: define only for nodes, not for all Products

prior to this change:
```
List((1,2,3)).toJson
// [{"head":1,"next":{"head":2,"next":{"head":3,"next":{}}}}]
```

now:
```
List((1,2,3)).toJson
// [1,2,3]
```

It's a regression that was introduced in https://github.com/joernio/joern/pull/5530
I have no idea why Michael Flanders made this change.
This is also adding a test so we catch it early next time round.

* fixup json serializer and tests

We don't want to serialize _all_ products, that's the main point of this
PR. But since Location is not a node and we still want to serialize it,
and it makes sense to have the product accessors, I added a marker trait
that let's us specifically format types using the product accessors...