[pysrc2cpg] Avoid tmp variable creation in for non complex receivers.
So far the python frontend emitted tmp variables in the context of call
site lowering if the `<x>` in a `<x>.y()` call was anything other than
an identifier/name. Now we also avoid emitting tmp variables if `<x>` is
a member access chain. E.g. `x.y.z()`. This leads to less clutter and
better alias handling.